Fix most sliders

+++ gl_bindings/src/opengl.rs
@@ -93,7 +93,7 @@ impl AbstractContext for GLContext {
     const TEXTURE_3D: u32 = gl::TEXTURE_3D;
     const UNSIGNED_BYTE: u32 = gl::UNSIGNED_BYTE;
     const RGBA: u32 = gl::RGBA;
-    const RGBA32F: u32 = gl::RGBA16F;
+    const RGBA32F: u32 = gl::RGBA32F;
     const LUMINANCE: u32 = gl::RED;
     const TEXTURE0: u32 = gl::TEXTURE0;
     const TEXTURE_WRAP_S: u32 = gl::TEXTURE_WRAP_S;
